Clean up leftover code
