+++ swervelib/abstract/__init__.py
@@ -10,10 +10,11 @@ __all__ = [
 from abc import ABCMeta
 from wpiutil import Sendable
 
-from motor import CoaxialDriveComponent, CoaxialAzimuthComponent
-from sensor import Gyro, AbsoluteEncoder
-from system import SwerveModule
-
 
 class SendableABCMeta(ABCMeta, type(Sendable)):
     pass
+
+
+from .motor import CoaxialDriveComponent, CoaxialAzimuthComponent
+from .sensor import Gyro, AbsoluteEncoder
+from .system import SwerveModule
